+++ app/src/main/java/com/example/cmput301f22t13/domainlayer/item/IngredientItem.java
@@ -1,18 +1,13 @@
 package com.example.cmput301f22t13.domainlayer.item;
 
-import com.example.cmput301f22t13.domainlayer.utils.Utils;
 import com.google.firebase.Timestamp;
 
 import java.io.Serializable;
 import java.util.Date;
 import java.util.GregorianCalendar;
-import java.util.Locale;
-import java.util.TimeZone;
-import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.Comparator;
 
 /** Public class representing an ingredient - options for constructing, getting and setting
 *
@@ -21,7 +16,7 @@ import java.util.Comparator;
 public class IngredientItem extends Item implements Serializable {
 
     private String description;
-    private Double amount;
+    private Integer amount;
     private String unit;
     private String category;
     private GregorianCalendar bbd;
@@ -32,7 +27,7 @@ public class IngredientItem extends Item implements Serializable {
      */
     public IngredientItem () {
         this.description = "";
-        this.amount = 0.0;
+        this.amount = 0;
         this.unit = "";
         this.category = "";
         this.bbd = new GregorianCalendar();
@@ -51,7 +46,7 @@ public class IngredientItem extends Item implements Serializable {
      * @param photo the uri string for storing images for the ingredient
      * @param location the location of the ingredient stored
      */
-    public IngredientItem (String name, String description, Double amount, String unit, String category, GregorianCalendar bbd, String photo, String location) {
+    public IngredientItem (String name, String description, Integer amount, String unit, String category, GregorianCalendar bbd, String photo, String location) {
         super(name, photo);
         this.description = description;
         this.amount = amount;
@@ -105,7 +100,7 @@ public class IngredientItem extends Item implements Serializable {
      * Gets the amount/number of units there are of this ingredient
      * @return amount
      */
-    public Double getAmount() {
+    public Integer getAmount() {
         return amount;
     }
 
@@ -113,7 +108,7 @@ public class IngredientItem extends Item implements Serializable {
      * Sets the amount/number of units there are of this ingredient
      * @param amount new amount
      */
-    public void setAmount(Double amount) {
+    public void setAmount(Integer amount) {
         this.amount = amount;
     }
